MongoDB connection
There's now a working API in Next.js, allowing me to completely replace the Flask API at any point if I so choose. Will likely get rid of it and practice with Flask in another project with Django.

Now the state is effectively managed between page reloads and the user's tracked tickers are pulled from the server on load.  The getTickerData function is currently unused, will need to re-evaluate whether it is necessary at all soon.