Effective DB communications
Reuses connections to the database and allows them to expire after extended periods of time, rather than overwhelming the server and continually opening new connections like the old method did.

NOTE: This method *does* need more testing, but is solid enough to implement for the time being as it still works.